Mixin working for pan-item and side-menu.  Allow's passing in selector for panElement
