+++ app/components/side-menu-button.js
@@ -3,7 +3,6 @@ import Ember from 'ember';
 export default Ember.Component.extend({
 
   click: function() {
-    console.log("SDFSdf");
-    this.sendAction('action', '123');
+    this.sendAction('action');
   }
 });
