+++ app/index.html
@@ -16,7 +16,7 @@
 
     <meta name="msapplication-tap-highlight" content="no" />
   </head>
-  <body>
+  <body class="no-select">
     {{content-for 'body'}}
 
     <script src="assets/vendor.js"></script>
