+++ app/templates/components/side-menu.hbs
@@ -1,3 +1,3 @@
-<div class="side-menu-button">X</div>
+<div class="side-menu-button" action="toggleMenu">X</div>
 
 {{yield}}
