+++ app/templates/list.hbs
@@ -13,16 +13,15 @@
     </pan-item>
 
     <div class="list-item-option">
-        {{fa-icon "camera" size="lg"}}
-        {{fa-icon "edit" size="lg"}}
-        {{fa-icon "envelope" size="lg"}}
-        {{fa-icon "flag-o" size="lg"}}  
-        {{fa-icon "heart" size="lg"}}
-        {{fa-icon "comment-o" size="lg"}}
-        {{fa-icon "cloud-download" size="lg"}}
-
+      {{fa-icon "camera" size="lg"}}
+      {{fa-icon "edit" size="lg"}}
+      {{fa-icon "envelope" size="lg"}}
+      {{fa-icon "flag-o" size="lg"}}  
+      {{fa-icon "heart" size="lg"}}
+      {{fa-icon "comment-o" size="lg"}}
+      {{fa-icon "cloud-download" size="lg"}}
     </div>
-
   </div>
+
 {{/each}}
 </div>
